refactor: reorganize tools and system libraries
