+++ src/09-tool-patterns/02-artsy-tools.ts
@@ -13,11 +13,9 @@ import chalk from "chalk"
 import { CoreMessage, generateText } from "ai"
 import { openai } from "@ai-sdk/openai"
 import dedent from "dedent"
-import {
-  fetchShowsNearLocation,
-  fetchUser,
-  fetchUserArtistFollows,
-} from "../lib/tools"
+import { fetchUser } from "@/tools/fetchUser"
+import { fetchUserArtistFollows } from "@/tools/fetchUserArtistFollows"
+import { fetchShowsNearLocation } from "@/tools/fetchShowsNearLocation"
 
 dotenv.config()
