--- src/lib/tools/index.ts
@@ -1,3 +1,0 @@
-export * from "./fetchUser"
-export * from "./fetchUserArtistFollows"
-export * from "./fetchShowsNearLocation"
