+++ src/system/metaphysics.ts
